+++ python/python/tests/test_async_client.py
@@ -26,6 +26,7 @@ from glide.async_commands.core import (
     UpdateOptions,
 )
 from glide.async_commands.sorted_set import (
+    AggregationType,
     InfBound,
     LexBoundary,
     Limit,
@@ -1651,7 +1652,8 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(key1, range) == {"a": 1.0, "d": 4.0}
+        zremrangebylex_res = await redis_client.zrange_withscores(key1, range)
+        assert compare_maps(zremrangebylex_res, {"a": 1.0, "d": 4.0}) is True
 
         assert (
             await redis_client.zremrangebylex(key1, LexBoundary("d"), InfBound.POS_INF)
@@ -1813,6 +1815,179 @@ class TestCommands:
         assert await redis_client.set(key2, "value") == OK
         with pytest.raises(RequestError):
             await redis_client.zmscore(key2, ["one"])
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_zinterstore(self, redis_client: TRedisClient):
+        key1 = "{testKey}:1-" + get_random_string(10)
+        key2 = "{testKey}:2-" + get_random_string(10)
+        key3 = "{testKey}:3-" + get_random_string(10)
+        range = RangeByIndex(0, -1)
+        members_scores1 = {"one": 1.0, "two": 2.0}
+        members_scores2 = {"one": 1.5, "two": 2.5, "three": 3.5}
+
+        assert await redis_client.zadd(key1, members_scores1) == 2
+        assert await redis_client.zadd(key2, members_scores2) == 3
+
+        assert await redis_client.zinterstore(key3, [key1, key2]) == 2
+        zinterstore_map = await redis_client.zrange_withscores(key3, range)
+        expected_map = {
+            "one": 2.5,
+            "two": 4.5,
+        }
+        assert compare_maps(zinterstore_map, expected_map) is True
+
+        # Intersection results are aggregated by the MAX score of elements
+        assert (
+            await redis_client.zinterstore(key3, [key1, key2], AggregationType.MAX) == 2
+        )
+        zinterstore_map_max = await redis_client.zrange_withscores(key3, range)
+        expected_map_max = {
+            "one": 1.5,
+            "two": 2.5,
+        }
+        assert compare_maps(zinterstore_map_max, expected_map_max) is True
+
+        # Intersection results are aggregated by the MIN score of elements
+        assert (
+            await redis_client.zinterstore(key3, [key1, key2], AggregationType.MIN) == 2
+        )
+        zinterstore_map_min = await redis_client.zrange_withscores(key3, range)
+        expected_map_min = {
+            "one": 1.0,
+            "two": 2.0,
+        }
+        assert compare_maps(zinterstore_map_min, expected_map_min) is True
+
+        # Intersection results are aggregated by the SUM score of elements
+        assert (
+            await redis_client.zinterstore(key3, [key1, key2], AggregationType.SUM) == 2
+        )
+        zinterstore_map_sum = await redis_client.zrange_withscores(key3, range)
+        expected_map_sum = {
+            "one": 2.5,
+            "two": 4.5,
+        }
+        assert compare_maps(zinterstore_map_sum, expected_map_sum) is True
+
+        # Scores are multiplied by 2.0 for key1 and key2 during aggregation.
+        assert (
+            await redis_client.zinterstore(
+                key3, [(key1, 2.0), (key2, 2.0)], AggregationType.SUM
+            )
+            == 2
+        )
+        zinterstore_map_multiplied = await redis_client.zrange_withscores(key3, range)
+        expected_map_multiplied = {
+            "one": 5.0,
+            "two": 9.0,
+        }
+        assert compare_maps(zinterstore_map_multiplied, expected_map_multiplied) is True
+
+        assert (
+            await redis_client.zinterstore(key3, [key1, "{testKey}-non_existing_key"])
+            == 0
+        )
+
+        # Empty list check
+        with pytest.raises(RequestError) as e:
+            await redis_client.zinterstore("{xyz}", [])
+        assert "wrong number of arguments" in str(e)
+
+        # Cross slot query
+        if isinstance(redis_client, RedisClusterClient):
+            with pytest.raises(RequestError) as e:
+                await redis_client.zinterstore("{xyz}", ["{abc}", "{def}"])
+            assert "CrossSlot" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_zunionstore(self, redis_client: TRedisClient):
+        key1 = "{testKey}:1-" + get_random_string(10)
+        key2 = "{testKey}:2-" + get_random_string(10)
+        key3 = "{testKey}:3-" + get_random_string(10)
+        range = RangeByIndex(0, -1)
+        members_scores1 = {"one": 1.0, "two": 2.0}
+        members_scores2 = {"one": 1.5, "two": 2.5, "three": 3.5}
+
+        assert await redis_client.zadd(key1, members_scores1) == 2
+        assert await redis_client.zadd(key2, members_scores2) == 3
+
+        assert await redis_client.zunionstore(key3, [key1, key2]) == 3
+        zunionstore_map = await redis_client.zrange_withscores(key3, range)
+        expected_map = {
+            "one": 2.5,
+            "three": 3.5,
+            "two": 4.5,
+        }
+        assert compare_maps(zunionstore_map, expected_map) is True
+
+        # Intersection results are aggregated by the MAX score of elements
+        assert (
+            await redis_client.zunionstore(key3, [key1, key2], AggregationType.MAX) == 3
+        )
+        zunionstore_map_max = await redis_client.zrange_withscores(key3, range)
+        expected_map_max = {
+            "one": 1.5,
+            "two": 2.5,
+            "three": 3.5,
+        }
+        assert compare_maps(zunionstore_map_max, expected_map_max) is True
+
+        # Intersection results are aggregated by the MIN score of elements
+        assert (
+            await redis_client.zunionstore(key3, [key1, key2], AggregationType.MIN) == 3
+        )
+        zunionstore_map_min = await redis_client.zrange_withscores(key3, range)
+        expected_map_min = {
+            "one": 1.0,
+            "two": 2.0,
+            "three": 3.5,
+        }
+        assert compare_maps(zunionstore_map_min, expected_map_min) is True
+
+        # Intersection results are aggregated by the SUM score of elements
+        assert (
+            await redis_client.zunionstore(key3, [key1, key2], AggregationType.SUM) == 3
+        )
+        zunionstore_map_sum = await redis_client.zrange_withscores(key3, range)
+        expected_map_sum = {
+            "one": 2.5,
+            "three": 3.5,
+            "two": 4.5,
+        }
+        assert compare_maps(zunionstore_map_sum, expected_map_sum) is True
+
+        # Scores are multiplied by 2.0 for key1 and key2 during aggregation.
+        assert (
+            await redis_client.zunionstore(
+                key3, [(key1, 2.0), (key2, 2.0)], AggregationType.SUM
+            )
+            == 3
+        )
+        zunionstore_map = await redis_client.zrange_withscores(key3, range)
+        expected_map = {
+            "one": 5.0,
+            "three": 7.0,
+            "two": 9.0,
+        }
+        assert compare_maps(zunionstore_map, expected_map) is True
+
+        assert (
+            await redis_client.zunionstore(key3, [key1, "{testKey}-non_existing_key"])
+            == 2
+        )
+
+        # Empty list check
+        with pytest.raises(RequestError) as e:
+            await redis_client.zunionstore("{xyz}", [])
+        assert "wrong number of arguments" in str(e)
+
+        # Cross slot query
+        if isinstance(redis_client, RedisClusterClient):
+            with pytest.raises(RequestError) as e:
+                await redis_client.zunionstore("{xyz}", ["{abc}", "{def}"])
+            assert "CrossSlot" in str(e)
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
@@ -2138,9 +2313,10 @@ class TestCommands:
             await redis_client.zrangestore(destination, source, RangeByIndex(0, -1))
             == 3
         )
-        assert await redis_client.zrange_withscores(
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"one": 1.0, "two": 2.0, "three": 3.0}
+        )
+        assert compare_maps(zrange_res, {"one": 1.0, "two": 2.0, "three": 3.0}) is True
 
         # range from rank 0 to 1, from highest to lowest score
         assert (
@@ -2149,9 +2325,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"three": 3.0, "two": 2.0}
+        )
+        assert compare_maps(zrange_res, {"two": 2.0, "three": 3.0}) is True
 
         # incorrect range, as start > stop
         assert (
@@ -2203,9 +2381,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"one": 1.0, "two": 2.0}
+        )
+        assert compare_maps(zrange_res, {"one": 1.0, "two": 2.0}) is True
 
         # range from 1 (inclusive) to positive infinity
         assert (
@@ -2214,9 +2394,10 @@ class TestCommands:
             )
             == 3
         )
-        assert await redis_client.zrange_withscores(
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"one": 1.0, "two": 2.0, "three": 3.0}
+        )
+        assert compare_maps(zrange_res, {"one": 1.0, "two": 2.0, "three": 3.0}) is True
 
         # range from negative to positive infinity, limited to ranks 1 to 2
         assert (
@@ -2227,9 +2408,10 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"two": 2.0, "three": 3.0}
+        )
+        assert compare_maps(zrange_res, {"two": 2.0, "three": 3.0}) is True
 
         # range from positive to negative infinity reversed, limited to ranks 1 to 2
         assert (
@@ -2241,9 +2423,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"two": 2.0, "one": 1.0}
+        )
+        assert compare_maps(zrange_res, {"one": 1.0, "two": 2.0}) is True
 
         # incorrect range as start > stop
         assert (
@@ -2308,9 +2492,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"a": 1.0, "b": 2.0}
+        )
+        assert compare_maps(zrange_res, {"a": 1.0, "b": 2.0}) is True
 
         # range from "a" (inclusive) to positive infinity
         assert (
@@ -2319,9 +2505,11 @@ class TestCommands:
             )
             == 3
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"a": 1.0, "b": 2.0, "c": 3.0}
+        )
+        assert compare_maps(zrange_res, {"a": 1.0, "b": 2.0, "c": 3.0}) is True
 
         # range from negative to positive infinity, limited to ranks 1 to 2
         assert (
@@ -2332,9 +2520,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"b": 2.0, "c": 3.0}
+        )
+        assert compare_maps(zrange_res, {"b": 2.0, "c": 3.0}) is True
 
         # range from positive to negative infinity reversed, limited to ranks 1 to 2
         assert (
@@ -2346,9 +2536,11 @@ class TestCommands:
             )
             == 2
         )
-        assert await redis_client.zrange_withscores(
+
+        zrange_res = await redis_client.zrange_withscores(
             destination, RangeByIndex(0, -1)
-        ) == {"b": 2.0, "a": 1.0}
+        )
+        assert compare_maps(zrange_res, {"a": 1.0, "b": 2.0}) is True
 
         # incorrect range as start > stop
         assert (
@@ -2495,10 +2687,9 @@ class TestCommands:
         assert await redis_client.zadd(key3, member_scores3) == 4
 
         assert await redis_client.zdiffstore(key4, [key1, key2]) == 2
-        assert await redis_client.zrange_withscores(key4, RangeByIndex(0, -1)) == {
-            "one": 1.0,
-            "three": 3.0,
-        }
+
+        zrange_res = await redis_client.zrange_withscores(key4, RangeByIndex(0, -1))
+        assert compare_maps(zrange_res, {"one": 1.0, "three": 3.0}) is True
 
         assert await redis_client.zdiffstore(key4, [key3, key2, key1]) == 1
         assert await redis_client.zrange_withscores(key4, RangeByIndex(0, -1)) == {
@@ -2521,6 +2712,79 @@ class TestCommands:
             with pytest.raises(RequestError) as e:
                 await redis_client.zdiffstore("abc", ["zxy", "lkn"])
             assert "CrossSlot" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_zrandmember(self, redis_client: TRedisClient):
+        key = get_random_string(10)
+        string_key = get_random_string(10)
+        scores = {"one": 1, "two": 2}
+        assert await redis_client.zadd(key, scores) == 2
+
+        member = await redis_client.zrandmember(key)
+        assert member in scores
+        assert await redis_client.zrandmember("non_existing_key") is None
+
+        # key exists, but it is not a set
+        assert await redis_client.set(string_key, "value") == OK
+        with pytest.raises(RequestError):
+            await redis_client.zrandmember(string_key)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_zrandmember_count(self, redis_client: TRedisClient):
+        key = get_random_string(10)
+        string_key = get_random_string(10)
+        scores = {"one": 1, "two": 2}
+        assert await redis_client.zadd(key, scores) == 2
+
+        # unique values are expected as count is positive
+        members = await redis_client.zrandmember_count(key, 4)
+        assert len(members) == 2
+        assert set(members) == {"one", "two"}
+
+        # duplicate values are expected as count is negative
+        members = await redis_client.zrandmember_count(key, -4)
+        assert len(members) == 4
+        for member in members:
+            assert member in scores
+
+        assert await redis_client.zrandmember_count(key, 0) == []
+        assert await redis_client.zrandmember_count("non_existing_key", 0) == []
+
+        # key exists, but it is not a set
+        assert await redis_client.set(string_key, "value") == OK
+        with pytest.raises(RequestError):
+            await redis_client.zrandmember_count(string_key, 5)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_zrandmember_withscores(self, redis_client: TRedisClient):
+        key = get_random_string(10)
+        string_key = get_random_string(10)
+        scores = {"one": 1, "two": 2}
+        assert await redis_client.zadd(key, scores) == 2
+
+        # unique values are expected as count is positive
+        elements = await redis_client.zrandmember_withscores(key, 4)
+        assert len(elements) == 2
+
+        for member, score in elements:
+            assert scores[str(member)] == score
+
+        # duplicate values are expected as count is negative
+        elements = await redis_client.zrandmember_withscores(key, -4)
+        assert len(elements) == 4
+        for member, score in elements:
+            assert scores[str(member)] == score
+
+        assert await redis_client.zrandmember_withscores(key, 0) == []
+        assert await redis_client.zrandmember_withscores("non_existing_key", 0) == []
+
+        # key exists, but it is not a set
+        assert await redis_client.set(string_key, "value") == OK
+        with pytest.raises(RequestError):
+            await redis_client.zrandmember_withscores(string_key, 5)
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])